Readme updated and reorganized project
