+++ README.md
@@ -1,19 +1,18 @@
 # Meu Projeto FastAPI
 
 This project is a RESTful API built with FastAPI and PostgreSQL. It includes CRUD operations for managing registers, along with data validation and error handling. The API supports creating, reading, updating, and deleting register entries.
-## Estrutura do Projeto
 
 
 ## Table of Contents
 
-- [Features]()
+- **Features**
 
-- [Requirements]()
-- [Installation]()
-- [Configuration]()
-- [Usage]()
-- [Testing]()
-- [Project Structure]()
+- **Requirements**
+- **Installation**
+- **Configuration**
+- **Usage**
+- **Testing**
+- **Project Structure**
 
 ## Features
 - **Create Register**: Create a new register entry.
