+++ app/database/register/deleteRegisterQuery.py
