+++ app/database/register/getRegisterByIdQuery.py
