+++ app/database/register/updateRegisterQuery.py
